+++ cloudant-sync-datastore-core/src/main/java/com/cloudant/sync/documentstore/Database.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2016 IBM Corp. All rights reserved.
+ * Copyright © 2016, 2018 IBM Corp. All rights reserved.
  *
  * Original iOS version by  Jens Alfke, ported to Android by Marty Schoch
  * Copyright © 2012 Couchbase, Inc. All rights reserved.
@@ -73,7 +73,10 @@ public interface Database {
     File getPath();
 
     /**
-     * <p>Returns the current winning revision of a document.</p>
+     * <p>Returns the current winning revision of a document; or returns the given
+     * <a href="https://couchdb.readthedocs.io/en/stable/api/local.html">local document</a>,
+     * if {@code documentId} is prefixed with {@code _local/}.
+     * </p>
      *
      * <p>Previously deleted documents can be retrieved
      * (via tombstones, see {@link Database#delete(DocumentRevision)})
@@ -109,7 +112,9 @@ public interface Database {
 
     /**
      * <p>Returns whether this DocumentStore contains a particular revision of
-     * a document.</p>
+     * a document; or contains the given
+     * <a href="https://couchdb.readthedocs.io/en/stable/api/local.html">local document</a>,
+     * if {@code documentId} is prefixed with {@code _local/}.</p>
      *
      * <p>{@code true} will still be returned if the document is deleted.</p>
      *
@@ -122,7 +127,10 @@ public interface Database {
     boolean contains(String documentId, String revisionId) throws DocumentStoreException;
 
     /**
-     * <p>Returns whether this DocumentStore contains any revisions of a document.
+     * <p>Returns whether this DocumentStore contains any revisions of a document; or contains the
+     * given
+     * <a href="https://couchdb.readthedocs.io/en/stable/api/local.html">local document</a>,
+     * if {@code documentId} is prefixed with {@code _local/}.
      * </p>
      *
      * <p>{@code true} will still be returned if the document is deleted.</p>
@@ -195,6 +203,7 @@ public interface Database {
 
     /**
      * <p>Return the number of documents in the DocumentStore</p>
+     * <p><b>Note:</b> this excludes local documents.</p>
      *
      * @return number of non-deleted documents in DocumentStore
      * @throws DocumentStoreException if there was an error reading from the database.
@@ -253,7 +262,10 @@ public interface Database {
         throws ConflictException;
 
     /**
-     * <p>Adds a new document with body and attachments from <code>rev</code>.</p>
+     * <p>Adds a new document with body and attachments from <code>rev</code>; or creates or updates
+     * a <a href="https://couchdb.readthedocs.io/en/stable/api/local.html">local document</a>,
+     * if {@code documentId} is prefixed with {@code _local/}.
+     * </p>
      *
      * <p>If the ID in <code>rev</code> is null, the document's ID will be auto-generated,
      * and can be found by inspecting the returned {@code DocumentRevision}.</p>
@@ -284,6 +296,8 @@ public interface Database {
      * {@link com.cloudant.sync.event.notifications.DocumentUpdated DocumentUpdated}
      * event is posted on the event bus.</p>
      *
+     * <p><b<Note:</b> to update local documents, call {@link #create(DocumentRevision)}</p>
+     *
      * @param rev the {@link DocumentRevision} to be updated
      * @return a {@link DocumentRevision} - the updated document
      * @throws ConflictException if <code>rev</code> is not a current revision for this document
@@ -296,7 +310,9 @@ public interface Database {
             AttachmentException, DocumentStoreException, DocumentNotFoundException;
 
     /**
-     * <p>Deletes a document from the DocumentStore.</p>
+     * <p>Deletes a document from the DocumentStore; or delete a
+     * <a href="https://couchdb.readthedocs.io/en/stable/api/local.html">local document</a>,
+     * if {@code documentId} is prefixed with {@code _local/}.</p>
      *
      * <p>This operation leaves a "tombstone" for the deleted document, so that
      * future replication operations can successfully replicate the deletion.
@@ -331,6 +347,8 @@ public interface Database {
      * <p>This is equivalent to calling
      * {@link Database#delete(DocumentRevision)
      * delete} on all leaf revisions</p>
+     *
+     * <p><b<Note:</b> to delete local documents, call {@link #delete(DocumentRevision)}</p>
      *
      * @param id the ID of the document to delete leaf nodes for
      * @return a List of a {@link DocumentRevision}s - the deleted or "tombstone" documents