+++ CHANGES.md
@@ -1,3 +1,8 @@
+# 2.4.0 (Unreleased)
+- [NEW] `Database` methods `read`, `contains`, `create`, and `delete` now accept local
+  (non-replicating documents). These documents must have their document ID prefixed with `_local/`
+  and must have their revision ID set to `null` (where applicable).
+  
 # 2.3.0 (2018-08-14)
 - [NEW] Added API for specifying a list of document IDs in the filtered pull replicator.
 - [IMPROVED] Forced a TLS1.2 `SSLSocketFactory` where possible on Android API versions < 20 (it is